pyproject.toml
[project]
name = "amritakalolsavam-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "boto3>=1.42.51",
    "django>=5.0,<6.0",
    "django-cors-headers>=4.9.0",
    "django-storages>=1.14.6",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "google-auth>=2.48.0",
    "gspread>=6.2.1",
    "msal>=1.34.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-slugify>=8.0.4",
    "sendgrid>=6.12.5",
]
